+++ R/freq.R
@@ -21,7 +21,6 @@
 #' @param accidentals \code{NULL} or character, represent accidentals,
 #' \code{"flat"} or \code{"sharp"}.
 #' @param collapse logical, collapse result into a single string.
-#' @param ... other arguments passed to \code{transpose}, specifically
 #' \code{key} and \code{style}.
 #'
 #' @return numeric vector