+++ NEWS.md
@@ -1,3 +1,7 @@
+# tabr 0.4.2
+
+* Added utility functions for retrieving LilyPond version and installation directory and `tabr` LilyPond API details.
+
 # tabr 0.4.1
 
 * Fixed `pitch_freq` documentation error.
@@ -8,7 +12,7 @@
     * The `bar` argument to `phrase` (and associated functions) is now `NULL` by default, or character, rather than simply `TRUE` or `FALSE`. If a string is provided, it is interpreted as LilyPond bar notation. E.g., `bar = "|"` adds the LilyPond syntax `\bar "|"` to the end of a phrase. 
     * If only a bar check is desired, `TRUE` is still accepted and will insert a bar check only rather than a literal bar. `FALSE` is treated as `NULL` for completeness.
 * Handle extra whitespace in `pc` and `pn`.
-* Fixed bug where final note in a scale was droppped by `scale_note`.
+* Fixed bug where final note in a scale was dropped by `scale_note`.
 * Unit test updates.
 * Documentation updates.